requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
matplotlib>=3.7.0  # For plotting results
tqdm>=4.65.0       # For progress bars
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
flask>=3.0.0
flask-sse>=0.2.1
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
werkzeug>=3.0.1
psycopg2-binary>=2.9.9
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
matplotlib>=3.7.0
tqdm>=4.65.0
torch>=2.0.0
torchvision>=0.15.0
redis>=5.0.0